refactor: move utility functions to their own file
