+++ src/components/JaavierR/helper/passwordGenerator.js
@@ -1,50 +1,15 @@
 import words from '../data.json'
-
-const ABC_ARRAY = [
-	'a',
-	'b',
-	'c',
-	'd',
-	'e',
-	'f',
-	'g',
-	'h',
-	'i',
-	'j',
-	'k',
-	'l',
-	'm',
-	'n',
-	'o',
-	'p',
-	'q',
-	'r',
-	's',
-	't',
-	'u',
-	'v',
-	'w',
-	'x',
-	'y',
-	'z'
-]
-
-function getUpper() {
-	return String.fromCharCode(Math.floor(Math.random() * 26 + 65))
-}
-
-function getLower() {
-	return String.fromCharCode(Math.floor(Math.random() * 26 + 97))
-}
-
-function getNumber() {
-	return Math.floor(Math.random() * 10)
-}
-
-function getSymbol() {
-	const symbols = [33, 42, 45, 46, 64, 95]
-	return String.fromCharCode(symbols[Math.floor(Math.random() * symbols.length)])
-}
+import {
+	ABC_ARRAY,
+	getLower,
+	getNumber,
+	getSymbol,
+	getUpper,
+	makeRandomString,
+	randomOddIntFromInterval,
+	setMinMaxLength,
+	setSeparator
+} from './utils'
 
 export function randomPassword({ length = 8, includeNumbers, includeSymbols }) {
 	let password = ''
@@ -234,39 +199,4 @@ export function memorablePassword({
 	}
 
 	return [...wordsArray.join(setSeparator(separator))]
-}
-
-function setMinMaxLength({ length, min, max }) {
-	return length < min ? min : length > max ? max : length
-}
-
-function setSeparator(separator) {
-	if (separator === 'Hyphens') {
-		return '-'
-	} else if (separator === 'Spaces') {
-		return ' '
-	} else if (separator === 'Periods') {
-		return '.'
-	} else if (separator === 'Commas') {
-		return ','
-	} else if (separator === 'Underscores') {
-		return '_'
-	}
-}
-
-function randomIntFromInterval(min, max) {
-	return Math.floor(Math.random() * (max - min + 1) + min)
-}
-
-function randomOddIntFromInterval(min, max) {
-	if (min % 2 === 0) ++min
-	return min + 2 * randomIntFromInterval(0, (max - min) / 2)
-}
-
-function makeRandomString(arr) {
-	const stringLength = Math.random() < 0.5 ? 3 : 4
-	return [...arr]
-		.sort(() => 0.5 - Math.random())
-		.splice(0, stringLength)
-		.join('')
 }
